create the profile and sidebar components
The profile component will serve as a hub for all of the user's clothing items that they add.
The sidebar will just be a section of the view that displays the username and user image.

There is currently hardcoded static text in these section. I plan in implementing the
cards and styles in the future.